Extract project extensions to separate module
