+++ packages/brookjs-cli/src/index.ts
@@ -1,9 +1,10 @@
 import * as eslint from './eslint';
-import * as glob from './glob';
 import * as fs from './fs';
+import * as glob from './glob';
+import * as project from './project';
 
 export * from './services';
 export * from './cli';
 export { default as create } from './create';
 
-export { glob, eslint, fs };
+export { eslint, fs, glob, project };
