+++ packages/brookjs-cli/src/services/index.ts
@@ -1,5 +1,4 @@
-import * as project from './project';
 import * as prettier from './prettier';
 
-export { project, prettier };
+export { prettier };
 export { default as WebpackService } from './webpack';
